Moved file processing to separate source file for easier testing.
